:sparkles: feat: learning how to handle Exception in Spring Boot, it is really cool!

+++ src/main/java/com/example/restful_api/service/ProductService.java
@@ -1,5 +1,6 @@
 package com.example.restful_api.service;
 
+import com.example.restful_api.exception.ItemNotFoundException;
 import com.example.restful_api.model.Product;
 import com.example.restful_api.repository.ProductRepository;
 import lombok.RequiredArgsConstructor;
@@ -18,14 +19,20 @@ public class ProductService {
         return productRepository.findAll();
     }
 
-    public Optional<Product> findById(Long id){
-        return productRepository.findById(id);
+    public Product findById(Long id){
+        return productRepository.findById(id)
+                .orElseThrow(() -> new ItemNotFoundException("Product with id "+ id +" not found."));
     }
 
     public Product save(Product product){
         return productRepository.save(product);
     }
     public void deleteById(Long id){
+
+        if(!productRepository.existsById(id)){
+            throw new ItemNotFoundException("Product with id "+ id +" not found.");
+        }
+
         productRepository.deleteById(id);
     }
